requirements.txt
openai==1.99.2
chromadb==1.2.1
python-box==7.3.2
PyYAML==6.0.2
requests==2.32.4
tavily-python